+++ src/terms/TermTreeSmall.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="mx-2 mb-1" :class="{ 'ml-4': level > 0 }">
     <div class="mb-1 flex flex-wrap items-baseline gap-4">
-      <Term :data="parent" :options="[search, goto]" />
+      <Term :data="parent" :options="[search, goto]" :draggable="true" />
       <div
         v-if="parent.narrower.length"
         class="cursor-pointer"